% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeptimedata-package.R
\docType{package}
\name{deeptimedata-package}
\alias{deeptimedata}
\alias{deeptimedata-package}
\title{deeptimedata: Geologic Pattern Data from FGDC Used in 'deeptime'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Geologic pattern data from \url{https://ngmdb.usgs.gov/fgdc_gds/geolsymstd.php}. Access functions are provided in the accompanying package 'deeptime'.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://williamgearty.com/deeptimedata/}
  \item \url{https://github.com/willgearty/deeptimedata}
  \item Report bugs at \url{https://github.com/willgearty/deeptimedata/issues}
}

}
\author{
\strong{Maintainer}: William Gearty \email{willgearty@gmail.com} (\href{https://orcid.org/0000-0003-0076-3262}{ORCID}) [copyright holder]

}
\keyword{internal}
